requirements.txt
fastapi
uvicorn
openai
celery[redis]
pydantic
python-multipart
google-genai
python-dotenv
pdf2image

